+++ extensions/stats/event.ts
@@ -13,7 +13,7 @@ import { ICountEvent } from './types/stats.ts';
 export function countEvent(
     events: BasicEvent[],
     boost: ColorBoostEvent[],
-    environment: EnvironmentAllName = 'DefaultEnvironment'
+    environment: EnvironmentAllName = 'DefaultEnvironment',
 ): ICountEvent {
     const commonEvent = EventList[environment]?.[0] ?? EventList['DefaultEnvironment'][0];
     const eventCount: ICountEvent = {};